Rename and move files. Update function names
Renamed and moved files to better work with naming convention.
Merged the UART files into a single .c/.h files which are at top level.
Added starting function for the power supply.
