+++ msp/msp_exp_handler.c
@@ -6,7 +6,7 @@
  */
 
 #include "../msp/msp_exp.h"
-#include "i2c_comm_int.h"
+#include "msp_i2c_int.h"
 
 static unsigned char *send_data;
 static unsigned char send_data_payload[30] = "This is data, important data";
